requirements.txt
google-generativeai==0.1.0
py-cord==2.4.1
python-dotenv==1.0.0
